--- a/.editorconfig
+++ b/.editorconfig
@@ -16,3 +16,8 @@ indent_size = 4
 [**.{yml,yaml}]
 indent_style = spaces
 indent_size = 2
+
+# Use spaces in package.json.
+[package.json]
+indent_style = spaces
+indent_size = 2