--- a/index.js
+++ b/index.js
@@ -4,18 +4,23 @@ const countries = require('./countries.json')
 const normalize = require('normalize-for-search')
 const levenshtein = require('leven')
 
-const match = (name) => {
+const matchCountry = (name) => {
 	if (countries[name]) return countries[name]
 
 	const normalized = normalize(name)
 	if (countries[normalized]) return countries[normalized]
 
-	// levenshtein not necessary, a relict of an older version. used to tolerate distance 1, but since there is "gambia" vs "sambia" in german, this doesnt work out
-	for(let country in countries){
-		if(levenshtein(country, normalized) === 0) return countries[country]
+	let matches = 0
+	let match = null
+	for (let country in countries){
+		if (levenshtein(country, normalized) <= 1) {
+			if (matches > 0) return null
+			matches++
+			match = countries[country]
+		}
 	}
-	return null
+	return match
 }
 
 
-module.exports = match
+module.exports = matchCountry