--- a/package.json
+++ b/package.json
@@ -25,11 +25,13 @@
     "slug": "^0.9.1"
   },
   "devDependencies": {
+    "benchmark": "^2.1.4",
     "tape": "^4.6.3"
   },
   "scripts": {
     "test": "node test.js",
-    "prepublish": "npm test"
+    "prepublish": "npm test",
+    "benchmark": "node benchmark.js"
   },
   "engine": {
     "node": ">=4"