--- a/package.json
+++ b/package.json
@@ -21,8 +21,8 @@
     "countries.json"
   ],
   "dependencies": {
-    "fast-levenshtein": "^2.0.5",
-    "slug": "^0.9.1"
+    "leven": "^2.1.0",
+    "normalize-for-search": "^2.0.1"
   },
   "devDependencies": {
     "benchmark": "^2.1.4",