--- a/test.js
+++ b/test.js
@@ -1,7 +1,17 @@
 'use strict'
 
 const tape = require('tape')
+const normalize = require('normalize-for-search')
+
+const countries = require('./countries.json')
 const match = require('.')
+
+tape('countries.json has normalized keys', (t) => {
+	for (let country in countries) {
+		t.equal(country, normalize(country))
+	}
+	t.end()
+})
 
 tape('match-country-german', (t) => {
 	t.plan(8)