7adbb8c re-adds fuzzy matching.

the other commits are cleanups or performance optimisations. :zap:

before:

```
exact match "Frankreich" x 557 ops/sec ±0.76% (88 runs sampled)
fuzzy match "Großbritannein" x 149 ops/sec ±0.40% (79 runs sampled)
too many typos "Großbrltannein" x 150 ops/sec ±1.07% (80 runs sampled)
invalid input "Test" x 239 ops/sec ±0.93% (82 runs sampled)
```

after:

```
exact match "Frankreich" x 494,314 ops/sec ±0.61% (85 runs sampled)
fuzzy match "Großbritannin" x 3,666 ops/sec ±0.56% (92 runs sampled)
too many typos "Großbrltannein" x 3,439 ops/sec ±0.39% (91 runs sampled)
invalid input "Test" x 12,570 ops/sec ±0.88% (87 runs sampled)
```